Move plugin customization for better organization.
