+++ app/Filament/Admin/Resources/UserResource/RelationManagers/UserLogRelationManager.php
@@ -2,7 +2,7 @@
 
 namespace App\Filament\Admin\Resources\UserResource\RelationManagers;
 
-use App\Filament\Resources\CustomActivitylogResource;
+use App\Filament\Plugins\ActivityLog\CustomActivitylogResource;
 use Filament\Forms\Form;
 use Filament\Resources\RelationManagers\RelationManager;
 use Filament\Tables\Actions\ViewAction;
